requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.0.1
six==1.15.0
Werkzeug==1.0.1
